playing around with alternative format
